fix: database migrations
