+++ server/libs/infra/src/index.ts
@@ -1,2 +1,3 @@
+export * from './database.config';
 export * from './infra.config';
 export * from './infra.module';
