+++ server/libs/infra/src/infra.config.ts
@@ -1,54 +1,6 @@
-import { QueueName } from '@app/domain';
 import { BullModuleOptions } from '@nestjs/bull';
-import { DataSource } from 'typeorm';
-import { PostgresConnectionOptions } from 'typeorm/driver/postgres/PostgresConnectionOptions';
 import { ConfigurationOptions } from 'typesense/lib/Typesense/Configuration';
-import {
-  AlbumEntity,
-  APIKeyEntity,
-  AssetEntity,
-  DeviceInfoEntity,
-  SharedLinkEntity,
-  SmartInfoEntity,
-  SystemConfigEntity,
-  UserEntity,
-  UserTokenEntity,
-} from './entities';
-
-const url = process.env.DB_URL;
-const urlOrParts = url
-  ? { url }
-  : {
-      host: process.env.DB_HOSTNAME || 'localhost',
-      port: parseInt(process.env.DB_PORT || '5432'),
-      username: process.env.DB_USERNAME || 'postgres',
-      password: process.env.DB_PASSWORD || 'postgres',
-      database: process.env.DB_DATABASE_NAME || 'immich',
-    };
-
-export const databaseConfig: PostgresConnectionOptions = {
-  type: 'postgres',
-  entities: [__dirname + '/entities/*.entity.{js,ts}'],
-  synchronize: false,
-  migrations: [__dirname + '/migrations/*.{js,ts}'],
-  migrationsRun: true,
-  connectTimeoutMS: 10000, // 10 seconds
-  ...urlOrParts,
-};
-
-export const dataSource = new DataSource(databaseConfig);
-
-export const databaseEntities = [
-  AssetEntity,
-  AlbumEntity,
-  APIKeyEntity,
-  DeviceInfoEntity,
-  UserEntity,
-  SharedLinkEntity,
-  SmartInfoEntity,
-  SystemConfigEntity,
-  UserTokenEntity,
-];
+import { QueueName } from '../../domain/src';
 
 export const bullConfig: BullModuleOptions = {
   prefix: 'immich_bull',
